pyproject.toml
[project]
name = "court-enrollment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.116.1",
    "line-bot-sdk>=3.19.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0",
    "uvicorn>=0.35.0",
]
